pyproject.toml
[tool.poetry]
name = "nipsu"
version = "0.0.1"
description = "Cross-platform network packet sniffer"
authors = [
  "Mateusz Koczan <xmateuszkoczan@gmail.com>"
]
license = "MIT License"
readme = "README.md"

documentation = "https://koczanm.github.io/nipus"
homepage = "https://koczanm.github.io/nipsu"
repository = "https://github.com/koczanm/nipsu"

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Topic :: Internet",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: System :: Networking",
  "Topic :: System :: Networking :: Monitoring",
  "Typing :: Typed",
]

packages = [
  {include = "nipsu", from = "src"}
]

[tool.poetry.scripts]
nipsu = "nipsu.cli:app"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
typer = {extras = ["all"], version = "^0.7.0"}
pypcap = {git = "https://github.com/koczanm/pypcap.git"}
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^1.7.7"
black = "^22.10.0"
flake8 = "^5.0.4"
flake8-bugbear = "^22.10.27"
flake8-builtins = "^2.0.1"
flake8-comprehensions = "^3.10.1"
flake8-debugger = "^4.1.2"
flake8-eradicate = "^1.4.0"
flake8-logging-format = "^0.8.1"
isort = "^5.10.1"
mypy = "^0.991"
pep8-naming = "^0.13.2"
pre-commit = "^2.20.0"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pyupgrade = "^3.2.2"
ipython = "^8.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 119
src_paths = ["src", "tests"]

[tool.black]
line-length = 119
target-version = ["py38", "py39", "py310"]
include = "./pyi?$"

[tool.pytest.ini_options]
addopts = """\
  --cov nipsu \
  --cov tests \
  --cov-report term-missing \
  --no-cov-on-fail
"""

[tool.coverage.report]
fail_under = 80
exclude_lines = [
  "if TYPE_CHECKING:",
  "pragma: no cover",
]

[tool.mypy]
check_untyped_defs = true
disallow_any_unimported = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true

[[tool.mypy.overrides]]
module = ["pcap"]
ignore_missing_imports = true
